Fixed SignIn/SignUp flow, pushes to correct routes, navBar directs to Home, Profile, Login
